Rachel's project draft
This is a draft of my final project. I need some help with editing the
Bootstrap CSS - I’ve tried to do it but the changes didn’t go into
effect. I am looking to edit the color scheme from blue/teal to
orange/navy/red. also, I love the <hr> style, but want to change the
star glyphicon to the cutlery glyphicon; when I did that, it removed
the horizontal lines and put the cutlery in a different spot near the
text rather than the same spot the star currently is in.